predjeno sa double na float, jer biblioteka radi sa float vrednostima

+++ src/Entity.cpp
@@ -5,13 +5,13 @@ Entity::Entity(double x, double y, const sf::Texture& texture)
 {
     m_image.setPosition(x, y);
 }
-Entity::Entity(const Vector2d& position, const sf::Texture& texture)
+Entity::Entity(const Vector2f& position, const sf::Texture& texture)
     :m_position(position), m_image(texture)
 {
     m_image.setPosition(position.x(), position.y());
 }
 
-const Vector2d& Entity::position() const {
+const Vector2f& Entity::position() const {
     return m_position;
 }
 const sf::Sprite& Entity::image() const {
@@ -21,7 +21,7 @@ const sf::Sprite& Entity::image() const {
 void Entity::setPosition(double x, double y) {
     m_position.setBoth(x, y);
 }
-void Entity::setPosition(const Vector2d& position) {
+void Entity::setPosition(const Vector2f& position) {
     m_position.setBoth(position.x(), position.y());
 }
